Fix fuzzer analysis to determine when it is safe to reuse reductions.
This is essentially a cache invalidation bug due to missing the
following case in the reduction checking logic:

 * Complete launch
 * Read-only privilege
 * Non-injective projection functor

The select_reduction code had been written to assume that the cache
invalidation case would always involve a write privilege. And in that
case the old code would correctly ensure that launches would use
injective projection functors. But that only covers the case when
downgrading a reduction privilege, and did not correctly handle the
case where the fuzzer chooses to send a read-only operation through
the pipeline. Because read-only operations have relaxed semantics even
when doing index launches, we had to reflow the check to cover those
invariants explicitly.